# simplify size_struct