size_struct currently iterates through properties in order.
When computing the size, the order is unnecessary.

benchmark                 old ns/op     new ns/op     delta
BenchmarkSize-12          6620          6394          -3.41%
BenchmarkSizeBytes-12     608           529           -12.99%